Reference

ProblemReductions.AbstractProblem

ProblemReductions.AbstractReductionResult

ProblemReductions.BruteForce

ProblemReductions.Circuit

ProblemReductions.CircuitSAT

ProblemReductions.GridGraph

ProblemReductions.HyperGraph

ProblemReductions.LogicGadget

ProblemReductions.ReductionCircuitToSpinGlass

ProblemReductions.SpinGlass

ProblemReductions.StaticBitVector

ProblemReductions.StaticElementVector

ProblemReductions.TruthTable

ProblemReductions.UnitDiskGraph

ProblemReductions.evaluate

ProblemReductions.extract_solution

ProblemReductions.findbest

ProblemReductions.flavors

ProblemReductions.num_flavors

ProblemReductions.num_terms

ProblemReductions.num_variables

ProblemReductions.onehotv

ProblemReductions.parameter_type

ProblemReductions.parameters

ProblemReductions.reduceto

ProblemReductions.reduction_complexity

ProblemReductions.set_parameters

ProblemReductions.spinglass_energy

ProblemReductions.spinglass_gadget

ProblemReductions.target_problem

ProblemReductions.terms

ProblemReductions.truth_table

ProblemReductions.variables

ProblemReductions.@bv_str

ProblemReductions.@circuit

ProblemReductions.AbstractProblem — Type

AbstractProblem

The abstract base type of computational problems.

ProblemReductions.AbstractReductionResult — Type

abstract type AbstractReductionResult

The base type for a reduction result.

ProblemReductions.BruteForce — Type

BruteForce

A brute force method to find the best configuration of a problem.

ProblemReductions.Circuit — Type

struct Circuit

A circuit expression is a sequence of assignments.

Fields

exprs::Vector{Assignment}: The assignments in the circuit.

ProblemReductions.CircuitSAT — Type

struct CircuitSAT <: AbstractProblem

Circuit satisfiability problem, where the goal is to find an assignment that satisfies the circuit.

Fields

circuit::Circuit: The circuit expression in SSA form.

symbols::Vector{Symbol}: The variables in the circuit.

ProblemReductions.GridGraph — Type

struct GridGraph <: Graphs.AbstractGraph{Int64}

A grid graph is a graph in which the vertices are arranged in a grid and two vertices are connected by an edge if and only if they are adjacent in the grid.

Fields

grid::BitMatrix: a matrix of booleans, where true indicates the presence of an edge.

radius::Float64: the radius of the unit disk

ProblemReductions.HyperGraph — Type

struct HyperGraph <: Graphs.AbstractGraph{Int64}

A hypergraph is a generalization of a graph in which an edge can connect any number of vertices.

Fields

n::Int: the number of vertices

edges::Vector{Vector{Int}}: a vector of vectors of integers, where each vector represents a hyperedge connecting the vertices with the corresponding indices.

ProblemReductions.LogicGadget — Type

struct LogicGadget{PT<:AbstractProblem}

The logic gadget defined on an computational model.

Fields

problem::PT: the computational model, e.g., SpinGlass.

inputs::Vector{Int}: the input variables.

outputs::Vector{Int}: the output variables.

References

What are the cost function for NAND and NOR gates?

Nguyen, M.-T., Liu, J.-G., Wurtz, J., Lukin, M.D., Wang, S.-T., Pichler, H., 2023. Quantum Optimization with Arbitrary Connectivity Using Rydberg Atom Arrays. PRX Quantum 4, 010316.

ProblemReductions.ReductionCircuitToSpinGlass — Type

struct ReductionCircuitToSpinGlass{GT, T}

The reduction result of a circuit to a spin glass problem.

Fields

spinglass::SpinGlass{GT, T}: the spin glass problem.

variables::Vector{Int}: the variables in the spin glass problem.

ProblemReductions.SpinGlass — Type

struct SpinGlass{GT<:Graphs.AbstractGraph, T} <: AbstractProblem

SpinGlass(graph::AbstractGraph, J, h=zeros(nv(graph)))

The spin-glass problem.

Positional arguments

graph is a graph object.

weights are associated with the edges.

ProblemReductions.StaticBitVector — Type

StaticBitVector{N,C} = StaticElementVector{N,1,C}
StaticBitVector(x::AbstractVector)

Examples

julia> sb = StaticBitVector([1,0,0,1,1])
10011

julia> sb[3]
0x0000000000000000

julia> collect(Int, sb)
5-element Vector{Int64}:
 1
 0
 0
 1
 1

ProblemReductions.StaticElementVector — Type

StaticElementVector{N,S,C}
StaticElementVector(nflavor::Int, x::AbstractVector)

N is the length of vector, C is the size of storage in unit of UInt64, S is the stride defined as the log2(# of flavors). When the number of flavors is 2, it is a StaticBitVector.

Fields

data is a tuple of UInt64 for storing the configuration of static elements.

Examples

julia> ev = StaticElementVector(3, [1,2,0,1,2])
12012

julia> ev[2]
0x0000000000000002

julia> collect(Int, ev)
5-element Vector{Int64}:
 1
 2
 0
 1
 2

ProblemReductions.TruthTable — Type

struct TruthTable{N, T}

The truth table.

Fields

inputs::Vector{T}: The input values.

outputs::Vector{T}: The output values.

values::Vector{BitStr{N, Int}}: The truth table values.

Examples

julia> tt = TruthTable(['a', 'b'], ['c'], [bit"0", bit"0", bit"0", bit"1"])
┌───┬───┬───┐
│ a │ b │ c │
├───┼───┼───┤
│ 0 │ 0 │ 0 │
│ 1 │ 0 │ 0 │
│ 0 │ 1 │ 0 │
│ 1 │ 1 │ 1 │
└───┴───┴───┘

ProblemReductions.UnitDiskGraph — Type

struct UnitDiskGraph{D, T} <: Graphs.AbstractGraph{Int64}

A unit disk graph is a graph in which the vertices are points in a plane and two vertices are connected by an edge if and only if the Euclidean distance between them is at most a given radius.

Fields

n::Int: the number of vertices

locations::Vector{NTuple{D, T}}: the locations of the vertices

radius::T: the radius of the unit disk

ProblemReductions.evaluate — Function

evaluate(problem::AbstractProblem, config) -> Real

Evaluate the energy of the problem given the configuration config. The lower the energy, the better the configuration.

ProblemReductions.extract_solution — Function

extract_solution(reduction::AbstractReductionResult, solution)

Extract the solution solution of the target problem to the original problem.

Arguments

reduction: The reduction result.

solution: The solution of the target problem.

ProblemReductions.findbest — Function

findbest(problem::AbstractProblem, method) -> Vector

Find the best configurations of the problem using the method.

ProblemReductions.flavors — Method

flavors(::Type{<:AbstractProblem}) -> Vector

Returns a vector of integers as the flavors (domain) of a degree of freedom.

ProblemReductions.num_flavors — Method

num_flavors(::Type{<:AbstractProblem}) -> Int

Returns the number of flavors (domain) of a degree of freedom.

ProblemReductions.num_terms — Method

num_terms(problem::AbstractProblem) -> Int

The number of terms in the computational problem.

ProblemReductions.num_variables — Method

num_variables(problem::AbstractProblem) -> Int

The number of variables in the computational problem.

ProblemReductions.onehotv — Method

onehotv(::Type{<:StaticElementVector}, i, v)
onehotv(::Type{<:StaticBitVector, i)

Returns a static element vector, with the value at location i being v (1 if not specified).

ProblemReductions.parameter_type — Method

parameter_type(problem::AbstractProblem) -> Type

The data type of the parameters in the computational problem.

ProblemReductions.parameters — Function

parameters(problem::AbstractProblem) -> Vector

Parameters associated with terms.

ProblemReductions.reduceto — Function

reduceto(::Type{TA}, x::AbstractProblem)

Reduce the problem x to a target problem of type TA. Returns an instance of AbstractReductionResult.

Arguments

TA: The target problem type.

x: The original problem.

ProblemReductions.reduction_complexity — Function

reduction_complexity(::Type{TA}, x::AbstractProblem) -> Int

The complexity of the reduction from the original problem to the target problem.

Arguments

TA: The target problem type.

x: The original problem.

ProblemReductions.set_parameters — Function

set_parameters(problem::AbstractProblem, parameters) -> AbstractProblem

Change the parameters for the problem and return a new problem instance.

ProblemReductions.spinglass_energy — Method

spinglass_energy(g::SimpleGraph, config; J, h)
spinglass_energy(cliques::AbstractVector{Vector{Int}}, config; weights)

Compute the spin glass state energy for the vertex configuration config. In the configuration, the spin ↑ is mapped to configuration 0, while spin ↓ is mapped to configuration 1. Let $G=(V,E)$ be the input graph, the hamiltonian is

\[H = \sum_{ij \in E} J_{ij} s_i s_j + \sum_{i \in V} h_i s_i,\]

where $s_i \in \{-1, 1\}$ stands for spin ↓ and spin ↑.

In the hypergraph case, the hamiltonian is

\[H = \sum_{c \in C} w_c \prod_{i \in c} s_i,\]

where $C$ is the set of cliques, and $w_c$ is the weight of the clique $c$.

ProblemReductions.spinglass_gadget — Method

spinglass_gadget(::Val{:arraymul})

The array multiplier gadget.

    s_{i+1,j-1}  p_i
           \     |
        q_j ------------ q_j
                 |
    c_{i,j} ------------ c_{i-1,j}
                 |     \
                 p_i     s_{i,j} 

variables: pi, qj, pq, c{i-1,j}, s{i+1,j-1}, c{i,j}, s{i,j}

constraints: 2 * c{i,j} + s{i,j} = pi qj + c{i-1,j} + s{i+1,j-1}

ProblemReductions.target_problem — Function

target_problem(res::AbstractReductionResult) -> AbstractProblem

Return the target problem of the reduction result.

ProblemReductions.terms — Function

terms(problem::AbstractProblem) -> Vector

The energy terms of a computational problem is defined as the variables that carrying local energies (or weights) in the computational problem.

ProblemReductions.truth_table — Method

truth_table(ga::LogicGadget; variables=1:num_variables(ga.problem), solver=BruteForce())

Compute the truth table of a logic gadget.

Arguments

ga::LogicGadget: the logic gadget.

Keyword Arguments

variables::Vector{Int}: the variables to be displayed.

solver::AbstractSolver: the solver to be used.

ProblemReductions.variables — Function

variables(problem::AbstractProblem) -> Vector

The degrees of freedoms in the computational problem. e.g. for the maximum independent set problems, they are the indices of vertices: 1, 2, 3..., while for the max cut problem, they are the edges.

ProblemReductions.@bv_str — Macro

Constructing a static bit vector.

ProblemReductions.@circuit — Macro

@circuit circuit_expr

Construct a circuit expression from a block of assignments.

Examples

julia> @circuit begin
        x = a ∨ b
        y = x ∧ c
       end
Circuit:
| x = ∨(a, b)
| y = ∧(x, c)