Reference

ProblemReductions.AbstractProblem

ProblemReductions.AbstractReductionResult

ProblemReductions.BoolVar

ProblemReductions.BruteForce

ProblemReductions.CNF

ProblemReductions.CNFClause

ProblemReductions.Circuit

ProblemReductions.CircuitSAT

ProblemReductions.Coloring

ProblemReductions.ConcatenatedReduction

ProblemReductions.DominatingSet

ProblemReductions.GridGraph

ProblemReductions.HyperGraph

ProblemReductions.IndependentSet

ProblemReductions.KSatisfiability

ProblemReductions.LogicGadget

ProblemReductions.MaxCut

ProblemReductions.QUBO

ProblemReductions.ReductionCircuitToSpinGlass

ProblemReductions.ReductionFactoringToSat

ProblemReductions.ReductionMaxCutToSpinGlass

ProblemReductions.ReductionQUBOToSpinGlass

ProblemReductions.ReductionSATTo3SAT

ProblemReductions.ReductionSATToIndependentSet

ProblemReductions.ReductionSatToColoring

ProblemReductions.ReductionSpinGlassToMaxCut

ProblemReductions.ReductionSpinGlassToQUBO

ProblemReductions.ReductionVertexCoveringToSetCovering

ProblemReductions.Satisfiability

ProblemReductions.SetPacking

ProblemReductions.SpinGlass

ProblemReductions.StaticBitVector

ProblemReductions.StaticElementVector

ProblemReductions.TruthTable

ProblemReductions.UnitDiskGraph

ProblemReductions.VertexCovering

ProblemReductions.:¬

ProblemReductions.:∧

ProblemReductions.:∨

ProblemReductions.configuration_space_size

ProblemReductions.evaluate

ProblemReductions.evaluate

ProblemReductions.evaluate

ProblemReductions.evaluate

ProblemReductions.evaluate

ProblemReductions.extract_solution

ProblemReductions.findbest

ProblemReductions.flavor_to_logical

ProblemReductions.flavors

ProblemReductions.is_set_covering

ProblemReductions.is_set_packing

ProblemReductions.is_vertex_coloring

ProblemReductions.is_vertex_covering

ProblemReductions.num_flavors

ProblemReductions.num_variables

ProblemReductions.onehotv

ProblemReductions.parameter_type

ProblemReductions.parameters

ProblemReductions.problem_size

ProblemReductions.problem_size

ProblemReductions.problem_size

ProblemReductions.problem_size

ProblemReductions.reduceto

ProblemReductions.reduction_complexity

ProblemReductions.satisfiable

ProblemReductions.set_parameters

ProblemReductions.spinglass_energy

ProblemReductions.spinglass_gadget

ProblemReductions.target_problem

ProblemReductions.truth_table

ProblemReductions.variables

ProblemReductions.@bools

ProblemReductions.@bv_str

ProblemReductions.@circuit

ProblemReductions.AbstractProblem — Type

AbstractProblem

The abstract base type of computational problems.

ProblemReductions.AbstractReductionResult — Type

abstract type AbstractReductionResult

The base type for a reduction result.

ProblemReductions.BoolVar — Type

BoolVar{T}
BoolVar(name, neg)

Boolean variable for constructing CNF clauses.

ProblemReductions.BruteForce — Type

struct BruteForce

A brute force method to find the best configuration of a problem.

ProblemReductions.CNF — Type

CNF{T}
CNF(clauses)

Boolean expression in conjunctive normal form. clauses is a vector of CNFClause, if and only if all clauses are satisfied, this CNF is satisfied.

Example

Under development

ProblemReductions.CNFClause — Type

CNFClause{T}
CNFClause(vars)

A clause in CNF, its value is the logical or of vars, where vars is a vector of BoolVar.

ProblemReductions.Circuit — Type

struct Circuit

A circuit expression is a sequence of assignments.

Fields

exprs::Vector{Assignment}: The assignments in the circuit.

ProblemReductions.CircuitSAT — Type

struct CircuitSAT <: AbstractProblem

Circuit satisfiability problem, where the goal is to find an assignment that satisfies the circuit.

Fields

circuit::Circuit: The circuit expression in SSA form.

symbols::Vector{Symbol}: The variables in the circuit.

ProblemReductions.Coloring — Type

struct Coloring{K, WT<:(AbstractVector)} <: AbstractProblem

Coloring{K}(graph; weights=UnitWeight(nv(graph)))

The Vertex Coloring problem.

Positional arguments

graph is the problem graph.

weights are associated with the edges of the graph, default to UnitWeight(ne(graph)).

ProblemReductions.ConcatenatedReduction — Type

struct ConcatenatedReduction

A sequence of reductions.

Fields

sequence::Vector{Any}: The sequence of reductions.

complexity::Int: The complexity of the reduction.

ProblemReductions.DominatingSet — Type

struct DominatingSet{GT<:Graphs.AbstractGraph} <: AbstractProblem

DominatingSet(graph; weights=UnitWeight())

The dominating set problem.

Positional arguments

graph is the problem graph.

We don't have weights for this problem.

ProblemReductions.GridGraph — Type

struct GridGraph <: Graphs.AbstractGraph{Int64}

A grid graph is a graph in which the vertices are arranged in a grid and two vertices are connected by an edge if and only if they are adjacent in the grid.

Fields

grid::BitMatrix: a matrix of booleans, where true indicates the presence of an edge.

radius::Float64: the radius of the unit disk

ProblemReductions.HyperGraph — Type

struct HyperGraph <: Graphs.AbstractGraph{Int64}

A hypergraph is a generalization of a graph in which an edge can connect any number of vertices.

Fields

n::Int: the number of vertices

edges::Vector{Vector{Int}}: a vector of vectors of integers, where each vector represents a hyperedge connecting the vertices with the corresponding indices.

ProblemReductions.IndependentSet — Type

struct IndependentSet{GT<:Graphs.AbstractGraph} <: AbstractProblem

The independent set problem in graph theory.

Positional arguments

graph is the problem graph.

This type of problem currently doesn't have weights.

Examples

Under Development

ProblemReductions.KSatisfiability — Type

struct KSatisfiability{K, T} <: ProblemReductions.AbstractSatisfiabilityProblem

The satisfiability problem for k-SAT, where the goal is to find an assignment that satisfies the CNF.

Fields

variables::Vector{T}: The variables in the CNF.

cnf::CNF{T}: The CNF expression.

ProblemReductions.LogicGadget — Type

struct LogicGadget{PT<:AbstractProblem}

The logic gadget defined on an computational model.

Fields

problem::PT: the computational model, e.g., SpinGlass.

inputs::Vector{Int}: the input variables.

outputs::Vector{Int}: the output variables.

References

What are the cost function for NAND and NOR gates?

Nguyen, M.-T., Liu, J.-G., Wurtz, J., Lukin, M.D., Wang, S.-T., Pichler, H., 2023. Quantum Optimization with Arbitrary Connectivity Using Rydberg Atom Arrays. PRX Quantum 4, 010316.

ProblemReductions.MaxCut — Type

struct MaxCut{WT<:(AbstractVector)} <: AbstractProblem

The cutting problem.

In this problem, we would like to find the cut of the graph that maximizes the sum of the

weights of the edges that are cut.

Positional arguments

graph is the problem graph.

weights are associated with the edges of the graph. We have ensure that the weights are in the same order as the edges in edges(graph).

ProblemReductions.QUBO — Type

struct QUBO{T<:Real} <: AbstractProblem

The quadratic unconstrained binary optimization.

\[E = \sum_{i,j} Q_{ij} x_i x_j\]

where x_i \in \{0, 1\}.

Arguments

matrix::AbstractMatrix: the matrix Q of the QUBO problem.

ProblemReductions.ReductionCircuitToSpinGlass — Type

struct ReductionCircuitToSpinGlass{GT, T}

The reduction result of a circuit to a spin glass problem.

Fields

num_source_vars::Int: the number of variables in the source circuit.

spinglass::SpinGlass{GT, T}: the spin glass problem.

variables::Vector{Int}: the variables in the spin glass problem.

ProblemReductions.ReductionFactoringToSat — Type

struct ReductionFactoringToSat <: AbstractReductionResult

The reduction result of a factoring problem to a CircuitSAT problem.

Fields

circuit::CircuitSAT: the CircuitSAT problem.

p::Vector{Int}: the first number to multiply (store bit locations)

q::Vector{Int}: the second number to multiply.

m::Vector{Int}: the result of the multiplication.

ProblemReductions.ReductionMaxCutToSpinGlass — Type

struct ReductionMaxCutToSpinGlass{GT, T}

The reduction result of a maxcut to a spin glass problem.

Fields

spinglass::SpinGlass{GT, T}: the spin glass problem.

We only consider a simple reduction from MaxCut to SpinGlass(the graph must be SimpleGraph).

ProblemReductions.ReductionQUBOToSpinGlass — Type

struct ReductionQUBOToSpinGlass{GT, T}

The reduction result of a qubo to a spin glass problem.

Fields

spinglass::SpinGlass{GT, T}: the spin glass problem.

We only consider a simple reduction from QUBO to SpinGlass(the graph must be SimpleGraph).

ProblemReductions.ReductionSATTo3SAT — Type

The reduction result of a general SAT problem to a 3-SAT problem.

Fields

sat_source::Satisfiability{GT, T}: the source general SAT problem.

ProblemReductions.ReductionSATToIndependentSet — Type

The reduction result of a general SAT problem to an Independent Set problem.

ProblemReductions.ReductionSatToColoring — Type

struct ReductionSatToColoring{K, T, WT<:(AbstractVector)}

The reduction result of a Sat problem to a Coloring problem.

Fields

Coloring{K, WT<:AbstractVector}: the coloring problem, where K is the number of colors and WT is the weights type.

varlabel, used to filter extra variables

Note: The coloring problem is a 3 coloring problem, in which a auxiliary color is used Auxiliary color => 2.

ProblemReductions.ReductionSpinGlassToMaxCut — Type

struct ReductionSpinGlassToMaxCut{WT}

The reduction result of a spin glass to a maxcut problem.

Fields

maxcut::MaxCut{WT}: the MaxCut problem.

ancilla::Int: the ancilla vertex.

ProblemReductions.ReductionSpinGlassToQUBO — Type

struct ReductionSpinGlassToQUBO{WT}

The reduction result of a spin glass to a QUBO problem.

Fields

qubo::QUBO{WT}: the QUBO problem.

ProblemReductions.ReductionVertexCoveringToSetCovering — Type

struct ReductionVertexCoveringToSetCovering{ET, WT<:(AbstractVector)}

The reduction result of a vertex covering to a set covering problem.

Fields

setcovering::SetCovering{ET,WT}: the set covering problem, where ET is the sets type and WT is the weights type.

edgelabel: map each edge to a number in order to identify the edge (otherwise the vector would be cluttering)

ProblemReductions.Satisfiability — Type

struct Satisfiability{T} <: ProblemReductions.AbstractSatisfiabilityProblem

The satisfiability problem.

Fields

cnf is a conjunctive normal form (CNF) for specifying the satisfiability problems.

weights are associated with clauses.

ProblemReductions.SetPacking — Type

struct SetPacking{ET} <: AbstractProblem

The set packing problem, a generalization of independent set problem to hypergraphs.

Positional arguments

elements is a vector of elements in the universe.

sets is a vector of vectors, each set is associated with a weight specified in weights.

Currently this type problem doesn't support weights.

ProblemReductions.SpinGlass — Type

struct SpinGlass{GT<:Graphs.AbstractGraph, WT<:(AbstractVector)} <: AbstractProblem

SpinGlass(graph::AbstractGraph, J, h=zeros(nv(graph)))

The spin-glass problem.

Positional arguments

graph is a graph object.

weights are associated with the edges.

ProblemReductions.StaticBitVector — Type

StaticBitVector{N,C} = StaticElementVector{N,1,C}
StaticBitVector(x::AbstractVector)

Examples

julia> sb = StaticBitVector([1,0,0,1,1])
10011

julia> sb[3]
0x0000000000000000

julia> collect(Int, sb)
5-element Vector{Int64}:
 1
 0
 0
 1
 1

ProblemReductions.StaticElementVector — Type

StaticElementVector{N,S,C}
StaticElementVector(nflavor::Int, x::AbstractVector)

N is the length of vector, C is the size of storage in unit of UInt64, S is the stride defined as the log2(# of flavors). When the number of flavors is 2, it is a StaticBitVector.

Fields

data is a tuple of UInt64 for storing the configuration of static elements.

Examples

julia> ev = StaticElementVector(3, [1,2,0,1,2])
12012

julia> ev[2]
0x0000000000000002

julia> collect(Int, ev)
5-element Vector{Int64}:
 1
 2
 0
 1
 2

ProblemReductions.TruthTable — Type

struct TruthTable{N, T}

The truth table.

Fields

inputs::Vector{T}: The input values.

outputs::Vector{T}: The output values.

values::Vector{BitStr{N, Int}}: The truth table values.

Examples

julia> tt = TruthTable(['a', 'b'], ['c'], [bit"0", bit"0", bit"0", bit"1"])
┌───┬───┬───┐
│ a │ b │ c │
├───┼───┼───┤
│ 0 │ 0 │ 0 │
│ 1 │ 0 │ 0 │
│ 0 │ 1 │ 0 │
│ 1 │ 1 │ 1 │
└───┴───┴───┘

ProblemReductions.UnitDiskGraph — Type

struct UnitDiskGraph{D, T} <: Graphs.AbstractGraph{Int64}

A unit disk graph is a graph in which the vertices are points in a plane and two vertices are connected by an edge if and only if the Euclidean distance between them is at most a given radius.

Fields

n::Int: the number of vertices

locations::Vector{NTuple{D, T}}: the locations of the vertices

radius::T: the radius of the unit disk

ProblemReductions.VertexCovering — Type

struct VertexCovering{WT<:(AbstractVector)} <: AbstractProblem

Vertex covering is a problem that seeks to find a minimum set of vertices that cover all edges in a graph.

Positional arguments

graph is a graph object.

weights are associated with the vertices of the graph, default to UnitWeight(nv(graph)).

ProblemReductions.:¬ — Method

¬(var::BoolVar)

Negation of a boolean variables of type BoolVar.

ProblemReductions.:∧ — Method

∧(vars...)

Logical and applied on CNFClause and CNF. Returns a new CNF.

ProblemReductions.:∨ — Method

∨(vars...)

Logical or applied on BoolVar and CNFClause. Returns a CNFClause.

ProblemReductions.configuration_space_size — Method

configuration_space_size(problem::AbstractProblem) -> Any

Return the log2 size of the configuration space of the problem.

ProblemReductions.evaluate — Function

evaluate(problem::AbstractProblem, config) -> Real

Evaluate the energy of the problem given the configuration config. The lower the energy, the better the configuration.

ProblemReductions.evaluate — Method

evaluate(c::Coloring, config)

Compute the energy of the vertex coloring configuration config, the energy is the number of violated edges.

ProblemReductions.evaluate — Method

evaluate(c::MaxCut, config)

Compute the cut weights for the vertex configuration config (an iterator). The energy is the sum of the weights of the edges that are cut.

ProblemReductions.evaluate — Method

evaluate(c::SetPacking, config)

First step: We check if config (a vector of boolean numbers as the mask of sets) is a set packing of sets;

Second step: If it is a set packing, we return - (size(set packing)); Otherwise, we return size(variables) + 1.

ProblemReductions.evaluate — Method

evaluate(c::VertexCovering, config)

return the weights of edge that is not covered but return typemax(eltype(weights)) if the edge is not covered. config is a vector of boolean numbers.

ProblemReductions.extract_solution — Function

extract_solution(reduction::AbstractReductionResult, solution)

Extract the solution solution of the target problem to the original problem.

Arguments

reduction: The reduction result.

solution: The solution of the target problem.

ProblemReductions.findbest — Function

findbest(problem::AbstractProblem, method) -> Vector

Find the best configurations of the problem using the method.

ProblemReductions.flavor_to_logical — Method

flavor_to_logical(::Type{T}, flavor) -> T

Convert the flavor to a logical value.

ProblemReductions.flavors — Method

flavors(::Type{<:AbstractProblem}) -> Vector

Returns a vector of integers as the flavors (domain) of a degree of freedom.

ProblemReductions.is_set_covering — Method

is_set_covering(c::SetCovering, config)

Return true if config (a vector of boolean numbers as the mask of sets) is a set covering of sets.

ProblemReductions.is_set_packing — Method

is_set_packing(sets::AbstractVector, config)

Return true if config (a vector of boolean numbers as the mask of sets) is a set packing of sets.

ProblemReductions.is_vertex_coloring — Method

is_vertex_coloring(graph::SimpleGraph, config)

Returns true if the coloring specified by config is a valid one, i.e. does not violate the contraints of vertices of an edges having different colors.

ProblemReductions.is_vertex_covering — Method

is_vertex_covering(graph::SimpleGraph, config)

return true if the vertex configuration config is a vertex covering of the graph. Our judgement is based on the fact that for each edge, at least one of its vertices is selected.

ProblemReductions.num_flavors — Method

num_flavors(::Type{<:AbstractProblem}) -> Int

Returns the number of flavors (domain) of a degree of freedom.

ProblemReductions.num_variables — Method

num_variables(problem::AbstractProblem) -> Int

The number of variables in the computational problem.

ProblemReductions.onehotv — Method

onehotv(::Type{<:StaticElementVector}, i, v)
onehotv(::Type{<:StaticBitVector, i)

Returns a static element vector, with the value at location i being v (1 if not specified).

ProblemReductions.parameter_type — Method

parameter_type(problem::AbstractProblem) -> Type

The data type of the parameters in the computational problem.

ProblemReductions.parameters — Function

parameters(problem::AbstractProblem) -> Vector

The parameters of the computational problem.

ProblemReductions.problem_size — Function

problem_size(problem::AbstractProblem) -> Int

The size of the computational problem. The measure of which depends on the problem.

ProblemReductions.problem_size — Method

problem size of the satisfiability problem is the number of clauses in the CNF.

ProblemReductions.problem_size — Method

Defined as the number of sets.

ProblemReductions.problem_size — Method

Defined as the number of elements times the number of sets.

ProblemReductions.reduceto — Function

reduceto(::Type{TA}, x::AbstractProblem)

Reduce the problem x to a target problem of type TA. Returns an instance of AbstractReductionResult.

Arguments

TA: The target problem type.

x: The original problem.

ProblemReductions.reduction_complexity — Method

reduction_complexity(::Type{TA}, x::AbstractProblem) -> Int

The complexity of the reduction from the original problem to the target problem. Returns the polynomial order of the reduction.

Note

The problem size measure is problem dependent. Please check problem_size for the problem size measure.

Arguments

TA: The target problem type.

x: The original problem.

ProblemReductions.satisfiable — Method

satisfiable(cnf::CNF, config::AbstractDict)

Returns true if an assignment of variables satisfies a CNF.

ProblemReductions.set_parameters — Function

set_parameters(problem::AbstractProblem, parameters) -> AbstractProblem

Change the parameters for the problem and return a new problem instance.

ProblemReductions.spinglass_energy — Method

spinglass_energy(g::SimpleGraph, config; J, h)
spinglass_energy(cliques::AbstractVector{Vector{Int}}, config; weights)

Compute the spin glass state energy for the vertex configuration config. In the configuration, the spin ↑ is mapped to configuration 0, while spin ↓ is mapped to configuration 1. Let $G=(V,E)$ be the input graph, the hamiltonian is

\[H = \sum_{ij \in E} J_{ij} s_i s_j + \sum_{i \in V} h_i s_i,\]

where $s_i \in \{-1, 1\}$ stands for spin ↓ and spin ↑.

In the hypergraph case, the hamiltonian is

\[H = \sum_{c \in C} w_c \prod_{i \in c} s_i,\]

where $C$ is the set of cliques, and $w_c$ is the weight of the clique $c$.

ProblemReductions.spinglass_gadget — Method

spinglass_gadget(::Val{:arraymul})

The array multiplier gadget.

    s_{i+1,j-1}  p_i
           \     |
        q_j ------------ q_j
                 |
    c_{i,j} ------------ c_{i-1,j}
                 |     \
                 p_i     s_{i,j} 

variables: pi, qj, pq, c{i-1,j}, s{i+1,j-1}, c{i,j}, s{i,j}

constraints: 2 * c{i,j} + s{i,j} = pi qj + c{i-1,j} + s{i+1,j-1}

ProblemReductions.target_problem — Function

target_problem(res::AbstractReductionResult) -> AbstractProblem

Return the target problem of the reduction result.

ProblemReductions.truth_table — Method

truth_table(ga::LogicGadget; variables=1:num_variables(ga.problem), solver=BruteForce())

Compute the truth table of a logic gadget.

Arguments

ga::LogicGadget: the logic gadget.

Keyword Arguments

variables::Vector{Int}: the variables to be displayed.

solver::AbstractSolver: the solver to be used.

ProblemReductions.variables — Function

variables(problem::AbstractProblem) -> Vector

The degrees of freedoms in the computational problem. e.g. for the maximum independent set problems, they are the indices of vertices: 1, 2, 3..., while for the max cut problem, they are the edges.

ProblemReductions.@bools — Macro

@bools(syms::Symbol...)

Create some boolean variables of type BoolVar in current scope that can be used in create a CNF.

Example

Under Development

ProblemReductions.@bv_str — Macro

Constructing a static bit vector.

ProblemReductions.@circuit — Macro

@circuit circuit_expr

Construct a circuit expression from a block of assignments.

Examples

julia> @circuit begin
        x = a ∨ b
        y = x ∧ c
       end
Circuit:
| x = ∨(a, b)
| y = ∧(x, c)